# perf(Expense): Optimize expense amounts stats query